Move the interactivity logic from r2r into a separate file
